requirements.txt
fastapi
pyyaml
uvicorn[standard]
jinja2
python-multipart
paramiko
git+https://github.com/XanaDublaKublaConch/sshcheck/